+++ app/src/main/java/com/nohjason/minari/MainActivity.kt
@@ -19,7 +19,6 @@ class MainActivity : ComponentActivity() {
                     modifier = Modifier.fillMaxSize(),
                     color = MaterialTheme.colorScheme.background
                 ) {
-//                    BottomNavGraph()
                     MainScreen()
                 }
             }
