+++ app/src/main/java/com/nohjason/minari/screens/HomeScreen.kt
@@ -1,18 +1,21 @@
 package com.nohjason.minari.screens
 
+import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.material3.Button
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
 
 @Composable
 fun HomeScreen() {
-    Column {
+    Column(
+        modifier = Modifier.fillMaxSize(),
+        horizontalAlignment = Alignment.CenterHorizontally,
+        verticalArrangement = Arrangement.Center
+    ) {
         Text(text = "This is Main Screen")
-        Button(onClick = {
-//            navController.navigate("profile")
-        }) {
-            Text(text = "go to Profile")
-        }
     }
 }
